All done, but what cost....

+++ Domain/Models/MaterialFile.cs
@@ -9,7 +9,7 @@ namespace Domain.Models
         public int IdFile { get; set; }
         public int? Volume { get; set; }
         public int Chapter { get; set; }
-        public int? FrameNumber { get; set; }
+        public byte? FrameNumber { get; set; }
 
         public virtual File IdFileNavigation { get; set; } = null!;
         public virtual Material IdMaterialNavigation { get; set; } = null!;
